Start drafting an update for audacity: 2.4.2 -> 3.0.2
The XGD-config patch is removed since it conflicts with the new version of audacity, as it was not merged upstream.

Audacity now requires their own fork of wxWidgets, according to https://wiki.audacityteam.org/wiki/Building_On_Linux - this is the cause of current trouble since I am new to nix and unsure how to integrate it. I copied the current 3.1.4 nixpkgs version of wxWidgets found at pkgs/development/libraries/wxwidgets/3.1/default.nix and modified it to retrieve the forked version, but that's it so far.