rustPlatform.buildRustPackage: do not add libiconv to buildInputs
Many buildRustPackage-based derivations require libiconv as a build
input on Darwin. This is caused by `libc` unconditionally linking
against libiconv. However, this was recently changed:

https://github.com/rust-lang/libc/pull/2150

We will probably see this dependency disappear over the coming months
once a new `libc` crate version is released and other crates adopt
this new version.

Since this is already an optional dependency and it will disappear
in the coming time, we should probably not unconditionally add it
to all Rust crates.