xfce.thunar: patch code execution vulnerability
When called with a regular file as command line argument, Thunar
would delegate to some other program without user confirmation
based on the file type. This could be exploited to trigger code
execution in a chain of vulnerabilities.

The patch for the vulnerability introduced a regression, which is why we
apply a second patch.

Both can be removed when thunar>=4.17.3.

https://www.openwall.com/lists/oss-security/2021/05/09/2