compiler-rt: Revert passing `COMPILER_RT_OS_DIR` and not symlinking libs
In 7869d1654517c028aa76fc1dedc9b5ac301a867a I got rid of the symlinking
by forcing `COMPILER_RT_OS_DIR` to always be the empty string. I thought
this was good because it just make compiler-rt be installed in a normal
way.

However, various LLVM tools expect the `COMPILER_RT_OS_DIR` to be set
normally, and fail to find things when they aren't in the expected lib
subdir.

Maybe it would be best to patch that too in the long term, but for now
we just undo this change.