CMake: update versioning infrastructure
Be smarter with versioning information: we don't
need to generate a whole build system to query
the version number set in CMake. Use script mode,
and print the version. Unfortunately, `project()`
isn't valid in script mode so we need to lift the
version into a variable and stick in some boilerplate
for printing the (full) version.