haskellPackages.monoid-extras: pin to < 0.6
The diagrams libraries universe still uses monoid-extras 0.5.1 and build
failures start cropping up when upgrading it. Since no other packages I
am aware of use this, we can pin it to 0.5.1 which still uses Option
instead of Maybe (the former of which is deprecated).

Patching this seems like a lot of work that'd break other stuff because
Option is seemingly part of the public API. Let's hope they sort all of
that out themselves before base 4.15 comes around in our main package
set.