nixos/tests/prosody: Ensure the TLS certificate is still valid
NixOS 20.09 has been out for so long that the certificate that is built
for the test is not longer valid. The build has never been reexecuted
since the last stdenv/openssl rebuild and is thus expired.

By setting the expiry time on +100y this should be valid enough for all
the use cases that might arise for 20.09. If you, in 2121, end up here
because the certificate is no longer valid: I am not sorry for whatever
it is you are doing. I would have hoped the world has either entirely
ran out of energy sources or this project to have been abandoned long
ago. If you want to get this to build again: Just add a whitespace to
the openssl invocation.