llvmPackages_{7-12} stdenvs: link all resource files from compiler-rt
Likely due to an oversight, after 7869d1654517c028aa76fc1dedc9b5ac301a867a
some of the resource files of compiler-rt are installed to $dev/include
instead of $out/share which breaks the assumption of multiple cc
wrappers in llvmPackages. To resolve this without causing a mass rebuild
on darwin, link the files from the wrong location into the correct place
in clang's resource root.

The proper fix for this is #123103 which however causes a darwin stdenv
rebuild and should go through staging first.