ultrastardx: 2020.4.0 -> 2021-04-03
Not a lot has changed since last update, a few fixes were pushed to
the master tree.
We take the liberty to update to master for now.

This updates the ffmpeg version from 2.x to 4.x, which seems to be
supported upstream when we bend the minor version restriction a
little.
A simple karaoke song played just fine in my test.