_1password-gui: 8.0.33-52 -> 8.0.34
This includes an overhaul of the install script, using patchelf to set
the interpreter and the libraries. This reflects a lot of what I've
learned about electron + nixos over the last few months.

This change is motivated by the fact that 1Password's version of
electron changed, probably leading to crashes that I've been seeing when
I try to update the Nix derivation to the latest versions.

Co-authored-by: Daniël de Kok <me@github.danieldk.eu>