symlinkJoin: Add support for multiple outputs
Previously, `symlinkJoin` only cared about the main output of its
inputs. This makes it possible to specify which outputs we want to
link through the `outputs` parameter; if it's customized,
`symlinkJoin` will produce multiple outputs, each one the result of
combining the corresponding outputs of the packages listed in `paths`.