nixos/doc/releases: Do not tag the release
I have seen a customer use the release tag as a pinned version
and the same can happen in flake input urls or similar input
configs, like `niv --branch`.
This results in a out of date version of the _all_ system
packages, putting users at risk of security vulnerabilities
and other issues that are addressed during the lifetime of a
Nixpkgs/NixOS release.

So clearly the tag poses a risk, but does it have a benefit?
I don't think so. It does not name the branch-off point, so we
don't need it for git operations. It does not represent the
best or canonical version of the release either, as further
fixes always occur after the release date. If it were the case,
we'd tag all channel updates, but we don't, so for the same
reasons, we should not tag the release.