treewide: replace createFindlibDestdir with equivalent preInstall hook
Resolves #121991. `createFindlibDestdir` has the flaw of running mkdir as
part of a setup hook which can fail in `nix-shell` when the store is
mounted read only. Additionally the `mkdir` is executed at an “improper”
time when it really should be done in `installPhase`.

Fixing the flaw in the feature of the setup hook would require to either
override `installPhase` somehow or `preInstall` which would both break
existing derivations and require extensive refactoring. Thus we instead
opt to remove this feature altogether since we won't be needing in the
future anyways (dune install doesn't require us to create
`$OCAMLFIND_DESTDIR`).