ocamlPackages.findlib: fail if createFindlibDestdir is used
This notifies users of this feature that is has been removed with a
descriptive error message as opposed to installPhase failing with an
error message that was previously fixed by setting `createFindlibDestdir
= true`.

See #121991 for the motivation for removing this. In short running `mkdir`
as part of the shell hook fails if the store is mounted read only (as it
usually is in `nix-shell`). Additionally we risk writing to store when
we shouldn't (if the store happens to be mounted writable).