postgresql: support JIT compilation for version >= 11
PostgreSQL (from 11 onwards AFAICT) supports JIT compilation of
(frequently used) queries. This should provide a speedup to some queries
where operations such as comparing rows is a bottleneck.

For JIT support we have to build PostgreSQL with with the --with-llvm
config flag and pass both clang and llvm into the build environment.
PostPostgreSQL then calls llvm-config to obtain the required compiler
flags.