solanum: fix MOTD
Previously this defaulted to the default MOTD in the solanum source
tree, and I don't want my friends to laugh at me. Includes a patch to
the tests to ensure that the MOTD is actually set.

This replicates the fix done in #109705 (solanum is a fork of charybdis,
so they share fundamental logic for this).

Signed-off-by: Christine Dodrill <me@christine.website>
(cherry picked from commit b1fe9fab6fcfa4cdefe7370ba264b281e48de0f8)