bozohttpd: experimental darwin compat patch
As far as I can tell, libSystem on darwin provides a libm compatibility
symlink, but not a libcrypt one.

Prior art (eg. CPython) appears to use `stdenv.isDarwin` as a proxy for
`stdenv.hostPlatform.libc == "libSystem"`, but the latter also works in
cases where we're building on Darwin, but with eg. musl for some reason.