tests/knot: test RFC42-style config
This currently fails to validate/reformat the knot config, because the
config relies on a config file from ~outside the sandbox~
defining a key. I worked around the problem of files from outside the
sandbox by removing "include" from the config for validation, but that
blows up as soon as keys that are only defined outside the sandbox are
referenced.