firefox-bin: Avoid including both the wrapped and unwrapped version when using nix run command.
In order to make the man pages accessible, the previous code used
nix-support/propagated-user-env-packages. However this file is also used to set
the PATH when the application is executed with `nix run`, thus including the
wrapped and the wrappee in the environment.

Having the wrappee enumerated first in the environment caused `firefox` to
default to the wrappee, and as such not being able to find a proper GTK. This
was a source of failures while opening a file-picker.

This change removes the code to propagate the wrappe in the environment, as the
man pages are already linked in the wrapper output.

(cherry picked from commit efef092ba59259fb2654038bbd952baca5c8bab7)